requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djoser==2.1.0
mysqlclient==2.1.0  
Pillow==8.2.0  
django-filter==2.4.0  
python-decouple==3.5  
django-cors-headers==3.7.0  


